# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Plataforma Web MOOC Académica
  description: Backend monolito modular en Go para la plataforma de cursos masivos abiertos en línea.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Entorno Docker Local
paths:
...
  /api/v1/courses:
    get:
      summary: Listar catálogo público de cursos
      responses:
        '200':
          description: Lista de cursos disponibles
    post:
      summary: Crear nuevo curso (Profesor / Admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, summary, passing_score]
              properties:
                title:
                  type: string
                summary:
                  type: string
                passing_score:
                  type: number
      responses:
        '200':
          description: Curso y versión borrador inicial creados
  /api/v1/courses/{id}:
    get:
      summary: Obtener jerarquía del curso publicado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Estructura del curso publicado
  /api/v1/courses/{id}/unpublish:
    post:
      summary: Despublicar un curso para mantenimiento (Profesor autor / Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Curso despublicado exitosamente
  /api/v1/courses/{id}/create-draft:
    post:
      summary: Crear borrador de actualización a partir de versión publicada
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Nuevo borrador creado preservando IDs estables
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer